requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
loguru==0.7.2

# LangChain & OpenAI
langchain==0.2.15
langchain-openai==0.1.23
openai==1.43.0

# Embeddings & Vector DB
chromadb==0.5.5

# Loaders & parsing
pypdf==4.3.1
unstructured==0.15.1
markdown==3.6
beautifulsoup4==4.12.3

# Frontend
streamlit==1.38.0

# Optional: re-ranking（如需）
rank-bm25==0.2.2
